requirements.txt

fastapi==0.115.0
uvicorn==0.30.6
aiohttp==3.10.5
sqlalchemy==2.0.35
aiomysql==0.2.0
python-jose[cryptography]==3.3.0
pydantic==2.9.2
openai-whisper==20240930
soundfile==0.12.1
numpy==1.26.4
coqui-tts==0.27.2
bcrypt==4.1.3
passlib==1.7.4
torch==2.8.0
torchaudio==2.8.0
numba==0.60.0